fix lại request

+++ Back-End/src/main/java/com/example/gymmanagement/dto/request/SignUpMembershipRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class SignUpMembershipRequest {
-    private Long memberId;
-    private Long membershipId;
+    private Integer memberId;
+    private Integer membershipId;
     private String payment_method;
 }
